backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
google-cloud-vision==3.7.0
Pillow==10.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
slowapi==0.1.9
redis==5.0.1
hashlib-compat==1.0.1
